remove service layer

+++ duva/tests/common.rs
@@ -367,7 +367,7 @@ impl Client {
         Ok(line.trim().to_string())
     }
 
-    pub fn send_and_get_vec(&mut self, command: impl AsRef<[u8]>, mut cnt: u16) -> Vec<String> {
+    pub fn send_and_get_vec(&mut self, command: impl AsRef<[u8]>, mut cnt: u32) -> Vec<String> {
         self.send(command.as_ref()).unwrap();
 
         let mut res = vec![];
